+++ src/main/kotlin/dev/usbharu/hideout/service/ap/APService.kt
@@ -5,13 +5,10 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
 import dev.usbharu.hideout.domain.model.ActivityPubResponse
 import dev.usbharu.hideout.domain.model.ap.Follow
-import dev.usbharu.hideout.domain.model.job.*
 import dev.usbharu.hideout.exception.JsonParseException
 import dev.usbharu.hideout.service.ap.job.APReceiveFollowJobService
 import dev.usbharu.hideout.service.ap.job.ApNoteJobService
 import dev.usbharu.hideout.service.ap.job.ApReactionJobService
-import kjob.core.dsl.JobContextWithProps
-import kjob.core.job.JobProps
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Qualifier
@@ -21,8 +18,6 @@ interface APService {
     fun parseActivity(json: String): ActivityType
 
     suspend fun processActivity(json: String, type: ActivityType): ActivityPubResponse?
-
-    suspend fun <T : HideoutJob> processActivity(job: JobContextWithProps<T>, hideoutJob: HideoutJob)
 }
 
 enum class ActivityType {
@@ -233,28 +228,5 @@ class APServiceImpl(
         }
     }
 
-    @Suppress("REDUNDANT_ELSE_IN_WHEN")
-    override suspend fun <T : HideoutJob> processActivity(job: JobContextWithProps<T>, hideoutJob: HideoutJob) {
-        logger.debug("processActivity: ${hideoutJob.name}")
 
-        @Suppress("ElseCaseInsteadOfExhaustiveWhen")
-        // Springで作成されるプロキシの都合上パターンマッチングが壊れるので必須
-        when (hideoutJob) {
-            is ReceiveFollowJob -> {
-                apReceiveFollowJobService.receiveFollowJob(
-                    job.props as JobProps<ReceiveFollowJob>
-                )
-            }
-
-            is DeliverPostJob -> apNoteJobService.createNoteJob(job.props as JobProps<DeliverPostJob>)
-            is DeliverReactionJob -> apReactionJobService.reactionJob(job.props as JobProps<DeliverReactionJob>)
-            is DeliverRemoveReactionJob -> apReactionJobService.removeReactionJob(
-                job.props as JobProps<DeliverRemoveReactionJob>
-            )
-
-            else -> {
-                throw IllegalStateException("WTF")
-            }
-        }
-    }
 }
